fix: home screen
