+++ src/screens/NotFoundScreen/index.tsx
@@ -29,8 +29,8 @@ export default function NotFoundScreen({
   return (
     <View style={styles.container}>
       <Text style={styles.title}>This screen does not exist.</Text>
-      <TouchableOpacity
-        onPress={() => navigation.replace("Root")}
+      <TouchableOpacity accessibilityRole="button"
+        onPress={() => navigation.replace("Home")}
         style={styles.link}
       >
         <Text style={styles.linkText}>Go to home screen!</Text>
