+++ src/screens/auth/ExpiredLinkScreen/index.tsx
@@ -29,7 +29,7 @@ function ExpiredLinkScreen() {
     if (accountId) {
       sendEmail(accountId);
     } else {
-      navigateTo("CausesScreen");
+      navigateTo("Cause");
     }
   };
