+++ src/screens/auth/SignInScreen/index.tsx
@@ -42,12 +42,12 @@ function SignInScreen() {
           <View style={S.contentContainer}>
             <Text style={S.title}>{t("title")}</Text>
             <GoogleLogin
-              onContinue={() => navigateTo("CausesScreen")}
+              onContinue={() => navigateTo("Cause")}
               from="direct_flow"
             />
             {Platform.OS === "ios" && (
               <AppleLogin
-                onContinue={() => navigateTo("CausesScreen")}
+                onContinue={() => navigateTo("Cause")}
                 from="direct_flow"
               />
             )}
